python-client/requirements.txt
pyzmq==25.1.1